Introduce Midje as BDD library
This adds Midje as a dev-dependency, meaning it will be available for developmental testing, but will not be packaged with releases of the compiler. For the best experience with Midje, developers should aff `lein-midje` to their profile.

(Open `~/.lein/profiles.clj` and add `{:user {:plugins [[lein-midje "3.0.0"]]}}`.)

This commit also adds the first test, ensuring that whitespace is insignificant and that parse trees are not influenced by it.

To this end, a new folder is introduced, which is intended to hold all the TIP code used for testing, so as to easily weed it out when creating a release.