Test behavior instead of implementation details
The new test suite for the parser decouples itself from the format out of returned AST. This is closer to the intention of testing behavior rather than implementation details.

NOTE: This does not imply that we should refrain from ever testing the AST produced by the parser. The basic "does it work at all" smoke test suite is just not the place.